feat(MongoModel): :sparkles: adicionando validacao ao método delete

+++ src/models/MongoModel.ts
@@ -37,6 +37,7 @@ abstract class MongoModel<T> implements IModel<T> {
   }
 
   public async delete(_id: string): Promise<T | null> {
+    if (!isValidObjectId(_id)) throw Error(ErrorTypes.InvalidMongoId);
     return this._model.findOneAndDelete({ _id });
   }
 }
